@@ -3,7 +3,6 @@ package main
 import (
 	"bytes"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"path"
@@ -11,6 +10,7 @@ import (
 	"text/template"
 	"time"
 
+	"github.com/alecthomas/kingpin/v2"
 	"github.com/apex/log"
 	"github.com/apex/log/handlers/cli"
 	"github.com/client9/codegen/shell"
@@ -18,7 +18,6 @@ import (
 	"github.com/goreleaser/goreleaser/pkg/context"
 	"github.com/goreleaser/goreleaser/pkg/defaults"
 	"github.com/pkg/errors"
-	kingpin "gopkg.in/alecthomas/kingpin.v2"
 )
 
 // nolint: gochecknoglobals
@@ -29,7 +28,7 @@ var (
 )
 
 // given a template, and a config, generate shell script
-func makeShell(tplsrc string, cfg *config.Project) ([]byte, error) {
+func makeShell(tplsrc string, cfg *Project) ([]byte, error) {
 	// if we want to add a timestamp in the templates this
 	//  function will generate it
 	funcMap := template.FuncMap{
@@ -60,7 +59,7 @@ func makePlatform(goos, goarch, goarm string) string {
 
 // makePlatformBinaries returns a map from platforms to a slice of binaries
 // built for that platform.
-func makePlatformBinaries(cfg *config.Project) map[string][]string {
+func makePlatformBinaries(cfg *Project) map[string][]string {
 	platformBinaries := make(map[string][]string)
 	for _, build := range cfg.Builds {
 		ignore := make(map[string]bool)
@@ -95,7 +94,6 @@ func makePlatformBinaries(cfg *config.Project) map[string][]string {
 // conditionals will return an error
 //
 // {{ .Binary }} --->  [prefix]${BINARY}, etc.
-//
 func makeName(prefix, target string) (string, error) {
 	// armv6 is the default in the shell script
 	// so do not need special template condition for ARM
@@ -110,11 +108,11 @@ func makeName(prefix, target string) (string, error) {
 	target = strings.Replace(target, "{{ .Arm }}", "{{ .Arch }}", -1)
 
 	// otherwise if it contains a conditional, we can't (easily)
-	// translate that to bash.  Ask for bug report.
+	// translate that to bash. Ask for bug report.
 	if strings.Contains(target, "{{ if") ||
 		strings.Contains(target, "{{if") {
-		//nolint: lll
-		return "", fmt.Errorf("name_template %q contains unknown conditional or ARM format. Please file bug at https://github.com/goreleaser/godownloader", target)
+		// nolint: lll
+		log.Warnf("⚠️ It contains a conditional, we can't (easily) translate that to bash. name_template %q contains unknown conditional or ARM format.", target)
 	}
 
 	varmap := map[string]string{
@@ -134,7 +132,7 @@ func makeName(prefix, target string) (string, error) {
 	if err != nil {
 		return "", err
 	}
-	err = t.Execute(&out, varmap)
+	err = t.Option("missingkey=zero").Execute(&out, varmap)
 	return out.String(), err
 }
 
@@ -153,7 +151,7 @@ func normalizeRepo(repo string) string {
 	return repo
 }
 
-func loadURLs(path, configPath string) (*config.Project, error) {
+func loadURLs(path, configPath string) (*Project, error) {
 	for _, file := range []string{configPath, "goreleaser.yml", ".goreleaser.yml", "goreleaser.yaml", ".goreleaser.yaml"} {
 		if file == "" {
 			continue
@@ -171,7 +169,7 @@ func loadURLs(path, configPath string) (*config.Project, error) {
 	return nil, fmt.Errorf("could not fetch a goreleaser configuration file")
 }
 
-func loadURL(file string) (*config.Project, error) {
+func loadURL(file string) (*Project, error) {
 	// nolint: gosec
 	resp, err := http.Get(file)
 	if err != nil {
@@ -182,22 +180,45 @@ func loadURL(file string) (*config.Project, error) {
 		return nil, nil
 	}
 	p, err := config.LoadReader(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	// dirty hack, based on knowledge of template
+	// see https://github.com/octomation/go-tool/blob/84688c88ae35e07d368ae86fd9c8f564444045be/.goreleaser.yml#L4-L14
+	w := Project{Project: p}
+	for _, arch := range p.Archives {
+		w.Archive = arch
+		break
+	}
 
 	// to make errcheck happy
 	errc := resp.Body.Close()
 	if errc != nil {
 		return nil, errc
 	}
-	return &p, err
+	return &w, err
 }
 
-func loadFile(file string) (*config.Project, error) {
+func loadFile(file string) (*Project, error) {
 	p, err := config.Load(file)
-	return &p, err
+	if err != nil {
+		return nil, err
+	}
+
+	// dirty hack, based on knowledge of template
+	// see https://github.com/octomation/go-tool/blob/84688c88ae35e07d368ae86fd9c8f564444045be/.goreleaser.yml#L4-L14
+	w := Project{Project: p}
+	for _, arch := range p.Archives {
+		w.Archive = arch
+		break
+	}
+
+	return &w, err
 }
 
 // Load project configuration from a given repo name or filepath/url.
-func Load(repo, configPath, file string) (project *config.Project, err error) {
+func Load(repo, configPath, file string) (project *Project, err error) {
 	if repo == "" && file == "" {
 		return nil, fmt.Errorf("repo or file not specified")
 	}
@@ -225,14 +246,14 @@ func Load(repo, configPath, file string) (project *config.Project, err error) {
 		project.Release.GitHub.Name = path.Base(repo)
 	}
 
-	var ctx = context.New(*project)
+	var ctx = context.New(project.Project)
 	for _, defaulter := range defaults.Defaulters {
 		log.Infof("setting defaults for %s", defaulter)
 		if err := defaulter.Default(ctx); err != nil {
 			return nil, errors.Wrap(err, "failed to set defaults")
 		}
 	}
-	project = &ctx.Config
+	project.Project = ctx.Config
 
 	// set default binary name
 	if len(project.Builds) == 0 {
@@ -295,7 +316,7 @@ func main() {
 	// only write out if forced to, OR if output is effectively different
 	// than what the file has.
 	if *force || shell.ShouldWriteFile(*output, out) {
-		if err = ioutil.WriteFile(*output, out, 0666); err != nil {
+		if err = os.WriteFile(*output, out, 0666); err != nil {
 			log.WithError(err).Errorf("unable to write to %s", *output)
 			os.Exit(1)
 		}
